server/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
